làm trống hóa đơn khi tạo mới && hiện tên nhân viên và chức vụ && connection string = localhost

+++ DataTransferObject/NhanVien.cs
@@ -11,10 +11,10 @@ namespace DataTransferObject
         public string MaNhanVien { get; set; }
         public string TenNhanVien { get; set; }
         public DateTime NgaySinh { get; set; }
-        public short ChucVu { get; set; }
+        public int ChucVu { get; set; }
         public string MatKhau { get; set; }
         public NhanVien() { }
-        public NhanVien(string maNV, string tenNV, DateTime ngaySinh, short chucVu, string matKhau)
+        public NhanVien(string maNV, string tenNV, DateTime ngaySinh, int chucVu, string matKhau)
         {
             MaNhanVien = maNV;
             TenNhanVien = tenNV;
